Added 3d.h
3d.h contains vector3d , camera, and triangle class. One step closer to adding triangles and forming a mesh.
