+++ 3D Graph/x64/Debug/3D Graph.tlog/3D Graph.lastbuildstate
@@ -1,0 +1,2 @@
+#TargetFrameworkVersion=v4.0:PlatformToolSet=v141:EnableManagedIncrementalBuild=false:VCToolArchitecture=Native32Bit:WindowsTargetPlatformVersion=10.0.17134.0
+Debug|x64|C:\Users\Shreyansh Darshan\Documents\GitHub\3D-Graph\|
